cheri-compiler-explorer: Switch to Ubuntu 20.04 from OpenSUSE Tumbleweed
The latter no longer works on the version of Docker present on our
Ubuntu hosts (20.04 at least, likely also 18.04). Sticking with (a) a
stable release of a distribution and (b) the same distribution as we run
Docker on should prove more reliable, and if we need newer versions of
things from the fast-paced world of Node.JS then we can always use
something like nvm.