Add combined SDK+QEMU Docker image
This relies on the fact that cheribsd-qemu-<arch> just adds files to
qemu-cheri and that cheribsd-sdk-<arch> is also indirectly based on
qemu-cheri via cheri-sdk. Were both halves to install additional
packages we would have to take extra care and manually install one
half's packages again in the combined image.