Split compiler explorer build out of SDK build
The two are completely separate builds, and don't even share artifacts,
yet currently use the same Jenkinsfile. We want ctsrd/cheri-sdk to be
triggered on rebuilds of its base ctsrd/qemu-cheri image, but
ctsrd/cheri-compiler-explorer just uses a plain Ubuntu base so should
only be triggered on compiler changes (and probably should also be
triggered on CheriBSD changes in order to pick up new sysroots, though
in practice it's rare that those change in a meaningful way for this
use).