Use a directory layout compatible with cheribuild
The kernel needs to live in the rootfs, the disk image needs to live in
output-root and the binaries need to live in tools-root/sdk. We could
hack all of these up with a cheribuild.json file but it quickly gets
messy, and if people start to build their own software inside with
cheribuild then things are just going to get weird.

Note that we slightly lie and create a rootfs that isn't a full rootfs,
just the sysroot and kernel. If it turns out that having the full rootfs
is useful we can fully populate it later.