More fixes for new layout
COPY foo bar copies the contents of foo, not foo itself, when foo is a
directory. There's also no way to use the --strip-components=1 tar flag
with ADD so instead do a multi-stage build to unpack it separately and
get to copy it to a different directory without bloating the layers.