cheri-compiler-explorer: Explicitly set instructionSet properties
The mips and riscv64 Clang ones get inferred, as do the non-Morello
aarch64 Clang ones, but non-Clang ones (due to a lack of -target) and
non-Morello aarch64 ones (due to -march not mentioning aarch64) do not.

Note that these names are chosen based on what CE's set of known ones
are, hence the seemingly-inconsistent naming between mips and riscv64.